+++ gql_dedupe_link/lib/gql_dedupe_link.dart
@@ -2,7 +2,7 @@
 library gql_dedupe_link;
 
 import "package:async/async.dart";
-import "package:gql/execution.dart";
+import "package:gql_exec/gql_exec.dart";
 import "package:gql_link/gql_link.dart";
 
 /// A [Link] to deduplicate [Request]s