+++ gql_transform_link/example/gql_transform_link_example.dart
@@ -1,4 +1,4 @@
-import "package:gql/execution.dart";
+import "package:gql_exec/gql_exec.dart";
 import "package:gql_link/gql_link.dart";
 import "package:gql_transform_link/gql_transform_link.dart";
 
@@ -9,7 +9,7 @@ void main() {
     TransformLink(
       requestTransformer: (req) => req.withContextEntry(
         RequestExtensionsThunk(
-          (req2) => req2.operation.variables,
+          (req2) => req2.variables,
         ),
       ),
       responseTransformer: (resp) => resp.withContextEntry(