+++ gql_transform_link/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.3
+
+- upgrade `package:gql_exec` to v0.2.0
+
+## 0.1.2
+
+- use `package:gql_exec`
+
 ## 0.1.1
 
 - upgrade `package:gql_link` to v0.2.0